+++ components/src/completion/Completion.ts
@@ -100,6 +100,7 @@ export default class Completion extends RapidElement {
   static parser = new ExcellentParser('@', [
     'contact',
     'fields',
+    'globals',
     'urns',
   ]);
 
@@ -144,6 +145,9 @@ export default class Completion extends RapidElement {
   @property({type: String})
   fieldsEndpoint: string;
 
+  @property({type: String})
+  globalsEndpoint: string;
+
   @property({type: Boolean})
   textarea: boolean;
 
@@ -154,6 +158,7 @@ export default class Completion extends RapidElement {
   public firstUpdated(changedProperties: Map<string, any>) {
     this.textInputElement = this.shadowRoot.querySelector("rp-textinput") as TextInput;
     this.anchorElement = this.shadowRoot.querySelector("#anchor");
+    this.keyedAssets = {};
 
     // TODO: fetch these once per page, not once per control
     if (this.completionsEndpoint) {
@@ -170,8 +175,14 @@ export default class Completion extends RapidElement {
 
     if (this.fieldsEndpoint) {
       getAssets(this.fieldsEndpoint).then((assets: Asset[])=>{
-        this.keyedAssets = { fields: assets.map((asset: Asset)=> asset.key ) }
+        this.keyedAssets["fields"] = assets.map((asset: Asset)=> asset.key );
       });      
+    }
+
+    if (this.globalsEndpoint) {
+      getAssets(this.globalsEndpoint).then((assets: Asset[])=>{
+        this.keyedAssets["globals"] = assets.map((asset: Asset)=> asset.key );
+      });
     }
 
     // create our hidden container so it gets included in our host element's form